Refactoring Code for Better Responsiveness
This commit includes refactoring of code to improve the responsiveness of the website. Changes have been made to ensure that the website adapts seamlessly to different screen sizes and devices, resulting in a better user experience. These modifications will help the website to perform optimally on a variety of platforms, from desktop to mobile devices.